+++ src/features/auth/index.ts
@@ -1,9 +1,9 @@
 // src/features/auth/index.ts
 
-export { Login } from './Login';
-export { AuthCallback } from './AuthCallback';
-export { ResetPassword } from './ResetPassword';
-export { EmailVerificationPending } from './EmailVerificationPending';
-export { PendingApproval } from './PendingApproval';
-export { DeniedAccess } from './DeniedAccess';
-export { PermissionDenied } from './PermissionDenied';
+export { Login } from "./Login";
+export { AuthCallback } from "./AuthCallback";
+export { ResetPassword } from "./ResetPassword";
+export { EmailVerificationPending } from "./EmailVerificationPending";
+export { PendingApproval } from "./PendingApproval";
+export { DeniedAccess } from "./DeniedAccess";
+export { PermissionDenied } from "./PermissionDenied";